requirements.txt
flask
sqlalchemy
psycopg2
python-bitcoinrpc
monero
